Use DI in i3pyblocks.modules.datetime
Using dependency injection instead of module patching seems to make much
easier to test. This makes the mock easier to do and make the contracts
clearer.

This example is pretty silly but it already removed one dependency from
project (freezegun).

The API for this is provisory, so put it behing a _ to make sure it is
"private".