Reconnect to Pulse in case of error
After suspending Pulse may return an error while trying to update.
Probably the reason is because the sound card may not be ready yet.

So, in case of error, reconnect to Pulse instance.

Sadly, didn't find a way to write a test for this case. This module
probably needs a rewrite since it is very complex.