Add core.Modules.setup()
This method does the setup that before was done in core.Modules.start().
The advantage of using a separate method for this is that the children
classes don't need to explicitly call `super().start(queue)`, so there
is less changes for things to go wrong. Also this hides an
implementation detail.