Use Mock instead of MagicMock
While `MagicMock` is really convenient, it ends up mocking too much so
it is easier to shoot yourself on the foot.

With `Mock`, we only mock the interface of the object that we pass as
parameters. Another advantage (when using asynctest.Mock) is that it
automatically mocks coroutine functions/methods.