Remove Poetry, use setup.py
Poetry was making the managment of this project more difficult than it
should. Also, this project is really simple in dependencies, and it is
more of a library and less of an application.

So not pinning dependencies also should make it easier to find issues
when i.e.: some dependency is broken.