Fix deadlock in i3pyblocks.blocks.pulsectl
Instead of using the same instance of Pulse() to control get the updates
and control the volume, use separate ones. This is a workaround, but
until we either rework this module in a proper async or we control the
internal pulsectl event loop correctly, this is as far as we can go.