Remove dependency injection
While DI helped mocking some classes, there were also some
disadvantages:

- The way I am using DI by injecting the class at constructor was
  showing at the documentation generated by Sphinx (and it was quite
  ugly).
- The usage was inconsistent between the project.

So let's go back to the mock approach. But there is a better way than
using mocker, that is using `patch`. This brings almost all benefits of
the DI, the only possible issue is when moving the class or renaming the
imports. But this shouldn't happen too much.