Add i3pyblocks.blocks.base.PollingSyncBlock
**Breaking change**

This is the equivalent of PollingBlock, but made for cases when asyncio
is not supported by the block and the calls for the functions maybe
slow. The case used as an example here is
i3pyblocks.blocks.x11.CaffeineBlock, since each call of X11 can block
for an unspecified amount of time.

Also rename ExecutorBlock to SyncBlock. This new name is a much better
description of what this block does. It also introduces
`click_handler_sync()` and `signal_handler_sync()`, and renames the main
method from `run()` to `run_sync()`. With this, it is much easier to
create a correct block using a fully synchronous dependency.